requirements.txt
discord
python-dotenv
pymongo
langchain-groq
langchain-core
langchain